Imported all files required for the lab and our code for the tool in salt.
We import the docker files of all the images we will use in our topology, our
docker-compose.yml and python support code. The tool is implemented as scripts
for salt external pillars and states. We also provide all the data to be
imported in the sources of truth (netbox, peering-manager) to have an up and
running topology that you can use to test the tool.
In this commit we include a couple of ssh keypairs that will be used for all
connectivity in the topology.