Change handling of custom policies.
Previously if a custom policy file existed for a session, it generated an extra
(custom) term in the produced OUT-bound policy. Currently, the custom policy file
adds an extra custom policy to be evaluated in the beginning of the OUT-bound
policy chain for a session. This should be a more flexible approach to introduce
custom policies in individual sessions.